Rename module confManager to conf_manager
