+++ test/test_schema_app_blocks.py
@@ -4,8 +4,8 @@
 import pytest
 from ve_utils.utype import UType as Ut
 from jsonschema.exceptions import SchemaError, ValidationError
-from vemonitor_m8.confManager.schema_validate import SchemaValidate
-from vemonitor_m8.confManager.loader import Loader
+from vemonitor_m8.conf_manager.schema_validate import SchemaValidate
+from vemonitor_m8.conf_manager.loader import Loader
 from .schema_test_helper import SchemaTestHelper
