+++ test/test_schema_emoncms.py
@@ -2,8 +2,8 @@
 import pytest
 from ve_utils.utype import UType as Ut
 from jsonschema.exceptions import SchemaError
-from vemonitor_m8.confManager.schema_validate import SchemaValidate
-from vemonitor_m8.confManager.loader import Loader
+from vemonitor_m8.conf_manager.schema_validate import SchemaValidate
+from vemonitor_m8.conf_manager.loader import Loader
 from .schema_test_helper import SchemaTestHelper
