+++ test/test_schema_validate_selector.py
@@ -2,12 +2,9 @@
 import inspect
 from os import path as Opath
 import pytest
-from ve_utils.utype import UType as Ut
-from jsonschema.exceptions import SchemaError
 from jsonschema.exceptions import ValidationError
-from vemonitor_m8.confManager.shema_validate_selector import SchemaValidateSelector as jValid
-from vemonitor_m8.confManager.config_loader import ConfigLoader
-from vemonitor_m8.core.exceptions import SettingInvalidException
+from vemonitor_m8.conf_manager.shema_validate_selector import SchemaValidateSelector as jValid
+from vemonitor_m8.conf_manager.config_loader import ConfigLoader
 
 
 @pytest.fixture(name="helper_manager", scope="class")
