+++ vemonitor_m8/apps/emoncms_structure.py
@@ -7,8 +7,8 @@ from ve_utils.utype import UType as Ut
 from vemonitor_m8.api.emon_data_helper import EmoncmsHelper
 from vemonitor_m8.api.emoncms_set_api import EmoncmsSetApi
 from vemonitor_m8.models.emoncms.delta_structure import DeltaStructure
-from vemonitor_m8.confManager.loader import Loader
-from vemonitor_m8.confManager.schema_validate import SchemaValidate as sValid
+from vemonitor_m8.conf_manager.loader import Loader
+from vemonitor_m8.conf_manager.schema_validate import SchemaValidate as sValid
 
 __author__ = "Eli Serra"
 __copyright__ = "Copyright 2020, Eli Serra"
