+++ vemonitor_m8/conf_manager/confFiles/victronDeviceData.yaml
