+++ vemonitor_m8/conf_manager/config_loader_helper.py
@@ -22,7 +22,7 @@ import logging
 
 from typing import Optional
 from ve_utils.utype import UType as Ut
-from vemonitor_m8.confManager.shema_validate_selector import SchemaValidateSelector as jValid
+from vemonitor_m8.conf_manager.shema_validate_selector import SchemaValidateSelector as jValid
 from vemonitor_m8.core.exceptions import SettingInvalidException, NullSettingException
 
 logging.basicConfig()
