+++ vemonitor_m8/conf_manager/schemas/appBlocks_schema.json
