+++ vemonitor_m8/conf_manager/schemas/appConnectors_schema.json
