+++ vemonitor_m8/conf_manager/schemas/batteryBanks_schema.json
