+++ vemonitor_m8/conf_manager/schemas/data_structure_schema.json
