+++ vemonitor_m8/conf_manager/schemas/emoncms_schema.json
