+++ vemonitor_m8/models/settings/config_helper.py
@@ -5,7 +5,7 @@ Configuration model Helper class
 """
 from typing import Optional
 from ve_utils.utype import UType as Ut
-from vemonitor_m8.confManager.shema_validate_selector import SchemaValidateSelector as jValid
+from vemonitor_m8.conf_manager.shema_validate_selector import SchemaValidateSelector as jValid
 from vemonitor_m8.core.exceptions import SettingInvalidException
 from vemonitor_m8.models.settings.app_block_helper import AppBlockHelper
